backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pydantic-extra-types==2.1.0
psycopg2-binary==2.9.7
asyncpg==0.28.0
python-dotenv==1.0.0
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.24.1
alembic==1.12.0
twilio==8.4.0 